nrfx_twi: Increase tx and rx transfer length
In order to send and receive more than 255 bytes in one transaction
this patch increases the size of the length parameters such that
data buffers larger than 255 bytes can be transferred in a single
transaction.

Also, this patch increases the HW_TIMEOUT define by a multiple of 10
to make sure that large transactions are not stopped early by timeout.