nrfx_saadc: Add channel enable/disable helper functions
Purpose of this addition is to have SAADC channel enable and
disable functions that modify only the active channel count and
modify the PSELP/PSELN registers. This is not to address any
bugs; it is so that one can enable and disable the channels
without the need to reconfigure them (they can be configured once
and then simply enabled/disabled).