nrfx_nvmc: Prevent use of NRF_UICR symbol in non-secure targets
In SoCs with TrustZone support, only secure code can access UICR.
When a non-secure target is built, the NRF_UICR symbol is not even
defined.
This commit prevents the NVMC driver from referencing the NRF_UICR
symbol when the driver is compiled for a non-secure target.
Assertions that normally use this symbol for checking if a given
address is valid for an operation on flash will fail in non-secure
code when the supplied address belongs to the UICR region, what is
the expected behavior.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>