drivers: nfct: Return an error instead of silently failing
It is dangerous to silently do the wrong thing. So instead of
returning invalid values when nrfx_nfct_nfcid1_default_bytes_get is
invoked on the non-secure side we return an error.

Also, the FICR_NFC_TAGHEADER0_MFGID_Msk define is no longer
checked. It was unnecessary to check it as it doesn't make sense to
compile this file wihout it defined.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>